+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.FortniteApp"
         tools:targetApi="31">
         <activity
-            android:name="main.main.MainActivity"
+            android:name="main.presentation.main.MainActivity"
             android:exported="true"
             android:label="@string/app_name">
             <intent-filter>
