+++ app/src/main/kotlin/main/presentation/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package main.main
+package main.presentation.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
