+++ app/src/main/kotlin/main/presentation/profile/ProfileFragment.kt
@@ -1,12 +1,12 @@
-package main.profile
+package main.presentation.profile
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import com.example.fortniteapp.databinding.FrProfileBinding
 import main.common.base.BaseFragment
-import main.profile.interactions.ProfileAction
-import main.profile.interactions.ProfileEvent
+import main.presentation.profile.interactions.ProfileAction
+import main.presentation.profile.interactions.ProfileEvent
 
 class ProfileFragment :
     BaseFragment<FrProfileBinding, ProfileViewModel, ProfileViewModel.State, ProfileAction, ProfileEvent>() {
