+++ app/src/main/kotlin/main/presentation/profile/interactions/ProfileAction.kt
@@ -1,0 +1,5 @@
+package main.presentation.profile.interactions
+
+sealed interface ProfileAction {
+
+}
