+++ app/src/main/kotlin/main/presentation/profile/interactions/ProfileEvent.kt
@@ -1,0 +1,4 @@
+package main.presentation.profile.interactions
+
+sealed interface ProfileEvent {
+}
