+++ app/src/main/kotlin/main/presentation/shop/ShopFragment.kt
@@ -1,19 +1,18 @@
-package main.battlepass
+package main.presentation.shop
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.fortniteapp.databinding.FrBattlePassBinding
 import com.example.fortniteapp.databinding.FrShopBinding
 
-class BattlePassFragment: Fragment() {
+class ShopFragment: Fragment() {
 
-    private lateinit var binding: FrBattlePassBinding
+    private lateinit var binding: FrShopBinding
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        binding = FrBattlePassBinding.inflate(inflater, container, false)
+        binding = FrShopBinding.inflate(inflater, container, false)
         return binding.root
     }
 }
