--- app/src/main/kotlin/main/profile/interactions/ProfileAction.kt
@@ -1,5 +1,0 @@
-package main.profile.interactions
-
-sealed interface ProfileAction {
-
-}
