--- app/src/main/kotlin/main/profile/interactions/ProfileEvent.kt
@@ -1,4 +1,0 @@
-package main.profile.interactions
-
-sealed interface ProfileEvent {
-}
