+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,25 +7,25 @@
 
     <fragment
         android:id="@+id/profileFragment"
-        android:name="main.profile.ProfileFragment"
+        android:name="main.presentation.profile.ProfileFragment"
         android:label="Profile"
         tools:layout="@layout/fr_profile" />
 
     <fragment
         android:id="@+id/battlePassFragment"
-        android:name="main.battlepass.BattlePassFragment"
+        android:name="main.presentation.battlepass.BattlePassFragment"
         android:label="Battle Pass"
         tools:layout="@layout/fr_battle_pass" />
 
     <fragment
         android:id="@+id/shopFragment"
-        android:name="main.shop.ShopFragment"
+        android:name="main.presentation.shop.ShopFragment"
         android:label="Shop"
         tools:layout="@layout/fr_shop" />
 
     <fragment
         android:id="@+id/settingsFragment"
-        android:name="main.settings.SettingsFragment"
+        android:name="main.presentation.settings.SettingsFragment"
         android:label="Settings"
         tools:layout="@layout/fr_settings" />
 </navigation>
